:sparkles: feat: 컴포넌트 서버 클라이언트 함수 로직 분리
